+++ tests/test-cran/test-paths.R
@@ -54,9 +54,12 @@ assert('url_filename() returns the file names in URLs', {
 })
 
 assert('is_abs_path() recognizes absolute paths on Windows and *nix', {
-  (!is_abs_path('abc/def'))
-  (is_abs_path(if (.Platform$OS.type == 'windows') {
-    c('D:\\abc', '\\\\netdrive\\somewhere')
+  # a trailing slash must not change the result (#127)
+  (is_rel_path(c('abc/def', 'abc/def/', 'test/mydir/', 'foo__files/')))
+  # '~' is absolute on all platforms (path.expand() maps it to the home dir)
+  (is_abs_path('~/foo'))
+  (is_abs_path(if (is_windows()) {
+    c('D:\\abc', '\\\\netdrive\\somewhere', 'C:/foo', '/foo', '\\foo')
   } else '/abc/def'))
 })